+++ noir-projects/noir-protocol-circuits/crates/private-kernel-lib/src/common.nr
@@ -5,8 +5,7 @@ use dep::types::{
     kernel_circuit_public_inputs::PrivateKernelCircuitPublicInputsBuilder,
     membership_witness::NoteHashReadRequestMembershipWitness,
     private_circuit_public_inputs::PrivateCircuitPublicInputs,
-    private_kernel::private_call_data::PrivateCallData,
-    kernel_data::{PrivateKernelInnerData, PrivateKernelTailData},
+    private_kernel::private_call_data::PrivateCallData, kernel_data::{PrivateKernelInnerData},
     side_effect::{SideEffect, SideEffectLinkedToNoteHash}
 },
     address::{AztecAddress, EthAddress, PartialAddress, compute_initialization_hash},
@@ -94,13 +93,14 @@ pub fn initialize_end_values(
     public_inputs.constants = previous_kernel.public_inputs.constants;
     public_inputs.min_revertible_side_effect_counter = previous_kernel.public_inputs.min_revertible_side_effect_counter;
 
+    let start = previous_kernel.public_inputs.validation_requests;
+    public_inputs.validation_requests.note_hash_read_requests = array_to_bounded_vec(start.note_hash_read_requests);
+    public_inputs.validation_requests.nullifier_read_requests = array_to_bounded_vec(start.nullifier_read_requests);
+    public_inputs.validation_requests.nullifier_key_validation_requests = array_to_bounded_vec(start.nullifier_key_validation_requests);
+
     // Ensure the arrays are the same as previously, before we start pushing more data onto them in other
     // functions within this circuit:
     let start = previous_kernel.public_inputs.end;
-
-    public_inputs.end.note_hash_read_requests = array_to_bounded_vec(start.note_hash_read_requests);
-    public_inputs.end.nullifier_read_requests = array_to_bounded_vec(start.nullifier_read_requests);
-    public_inputs.end.nullifier_key_validation_requests = array_to_bounded_vec(start.nullifier_key_validation_requests);
 
     public_inputs.end.new_note_hashes = array_to_bounded_vec(start.new_note_hashes);
     public_inputs.end.new_nullifiers = array_to_bounded_vec(start.new_nullifiers);
@@ -196,14 +196,14 @@ pub fn update_end_values(
             )
         }
     }
-    public_inputs.end.note_hash_read_requests.extend_from_bounded_vec(siloed_read_requests);
+    public_inputs.validation_requests.note_hash_read_requests.extend_from_bounded_vec(siloed_read_requests);
 
     // Nullifier read request.
     let nullifier_read_requests = private_call_public_inputs.nullifier_read_requests;
     for i in 0..MAX_NULLIFIER_READ_REQUESTS_PER_CALL {
         let request = nullifier_read_requests[i];
         if !is_empty(request) {
-            public_inputs.end.nullifier_read_requests.push(request.to_context(storage_contract_address));
+            public_inputs.validation_requests.nullifier_read_requests.push(request.to_context(storage_contract_address));
         }
     }
 
@@ -212,7 +212,7 @@ pub fn update_end_values(
     for i in 0..MAX_NULLIFIER_KEY_VALIDATION_REQUESTS_PER_CALL {
         let request = nullifier_key_validation_requests[i];
         if !is_empty(request) {
-            public_inputs.end.nullifier_key_validation_requests.push(request.to_context(storage_contract_address));
+            public_inputs.validation_requests.nullifier_key_validation_requests.push(request.to_context(storage_contract_address));
         }
     }
