+++ noir-projects/noir-protocol-circuits/crates/private-kernel-lib/src/private_kernel_tail.nr
@@ -4,7 +4,7 @@ use dep::reset_kernel_lib::{NullifierReadRequestHints, reset_read_requests};
 use dep::types::{
     abis::{
     call_request::CallRequest, nullifier_key_validation_request::NullifierKeyValidationRequestContext,
-    kernel_data::{PrivateKernelInnerData, PrivateKernelTailData},
+    kernel_data::PrivateKernelInnerData,
     kernel_circuit_public_inputs::{PrivateKernelCircuitPublicInputsBuilder, PrivateKernelTailCircuitPublicInputs},
     side_effect::{SideEffect, SideEffectLinkedToNoteHash, Ordered}
 },
@@ -38,7 +38,7 @@ impl PrivateKernelTailCircuitPrivateInputs {
     }
 
     fn validate_nullifier_read_requests(self, public_inputs: &mut PrivateKernelCircuitPublicInputsBuilder) {
-        let requests = self.previous_kernel.public_inputs.end.nullifier_read_requests;
+        let requests = self.previous_kernel.public_inputs.validation_requests.nullifier_read_requests;
 
         let pending_nullifiers = self.previous_kernel.public_inputs.end.new_nullifiers;
 
@@ -46,7 +46,7 @@ impl PrivateKernelTailCircuitPrivateInputs {
 
         let nullifier_tree_root = public_inputs.constants.historical_header.state.partial.nullifier_tree.root;
 
-        public_inputs.end.nullifier_read_requests = reset_read_requests(
+        public_inputs.validation_requests.nullifier_read_requests = reset_read_requests(
             requests,
             pending_nullifiers,
             hints.read_request_statuses,
@@ -58,12 +58,12 @@ impl PrivateKernelTailCircuitPrivateInputs {
         // corresponding values are added to public inputs in nested executions.
         // But right now, all the request must be cleared in one go.
         assert(
-            public_inputs.end.nullifier_read_requests.len() == 0, "All nullifier read requests must be verified"
+            public_inputs.validation_requests.nullifier_read_requests.len() == 0, "All nullifier read requests must be verified"
         );
     }
 
     fn validate_nullifier_keys(self, public_inputs: &mut PrivateKernelCircuitPublicInputsBuilder) {
-        let requests = self.previous_kernel.public_inputs.end.nullifier_key_validation_requests;
+        let requests = self.previous_kernel.public_inputs.validation_requests.nullifier_key_validation_requests;
         for i in 0..MAX_NULLIFIER_KEY_VALIDATION_REQUESTS_PER_TX {
             let request = requests[i];
             if !is_empty(request) {
@@ -81,12 +81,12 @@ impl PrivateKernelTailCircuitPrivateInputs {
         }
 
         // Empty out nullifier key validation requests after verifying them.
-        public_inputs.end.nullifier_key_validation_requests = BoundedVec::new();
+        public_inputs.validation_requests.nullifier_key_validation_requests = BoundedVec::new();
     }
 
     fn match_reads_to_commitments(self, public_inputs: &mut PrivateKernelCircuitPublicInputsBuilder) {
         let new_note_hashes = public_inputs.end.new_note_hashes;
-        let read_requests = public_inputs.end.note_hash_read_requests;
+        let read_requests = public_inputs.validation_requests.note_hash_read_requests;
 
         // match reads to commitments from the previous call(s)
         for rr_idx in 0..MAX_NOTE_HASH_READ_REQUESTS_PER_TX {
@@ -103,7 +103,7 @@ impl PrivateKernelTailCircuitPrivateInputs {
         }
 
         // Empty out read requests after matching them to commitments
-        public_inputs.end.note_hash_read_requests = BoundedVec::new();
+        public_inputs.validation_requests.note_hash_read_requests = BoundedVec::new();
     }
 
     fn assert_sorted_counters<T, N>(original: [T; N], sorted: [T; N], indexes: [u64; N]) where T: Eq + Ordered + Empty {
@@ -447,8 +447,8 @@ mod tests {
         builder.append_transient_commitments(1);
         builder.add_transient_read(0);
         // Tweak the read request so that it does not match the hash at index 0;
-        let read_request = builder.previous_kernel.end.note_hash_read_requests.pop();
-        builder.previous_kernel.end.note_hash_read_requests.push(SideEffect { value: read_request.value + 1, counter: 0 });
+        let read_request = builder.previous_kernel.validation_requests.note_hash_read_requests.pop();
+        builder.previous_kernel.validation_requests.note_hash_read_requests.push(SideEffect { value: read_request.value + 1, counter: 0 });
         builder.failed();
     }
 
@@ -494,9 +494,9 @@ mod tests {
         builder.append_nullifiers(3);
         builder.add_nullifier_pending_read(1);
         let nullifier_being_read = builder.get_new_nullifiers()[2];
-        let mut read_request = builder.previous_kernel.end.nullifier_read_requests.pop();
+        let mut read_request = builder.previous_kernel.validation_requests.nullifier_read_requests.pop();
         read_request.counter = nullifier_being_read.counter - 1;
-        builder.previous_kernel.end.nullifier_read_requests.push(read_request);
+        builder.previous_kernel.validation_requests.nullifier_read_requests.push(read_request);
 
         builder.failed();
     }
