+++ noir-projects/noir-protocol-circuits/crates/public-kernel-lib/src/public_kernel_teardown.nr
@@ -43,6 +43,8 @@ impl PublicKernelTeardownCircuitPrivateInputs {
         let call_request = public_inputs.end_non_revertible.public_call_stack.pop();
         common::validate_call_against_request(self.public_call, call_request);
 
+        common::update_validation_requests(self.public_call, &mut public_inputs);
+
         common::update_public_end_non_revertible_values(self.public_call, &mut public_inputs);
 
         if public_inputs.reverted == false {
