+++ noir-projects/noir-protocol-circuits/crates/types/src/abis/accumulated_data/combined_accumulated_data_builder.nr
@@ -9,21 +9,14 @@ use crate::{
     public_accumulated_revertible_data::PublicAccumulatedRevertibleData,
     public_accumulated_non_revertible_data::PublicAccumulatedNonRevertibleData
 },
-    call_request::CallRequest, nullifier_key_validation_request::NullifierKeyValidationRequestContext,
-    public_data_read::PublicDataRead, public_data_update_request::PublicDataUpdateRequest,
-    read_request::ReadRequestContext, side_effect::{SideEffect, SideEffectLinkedToNoteHash}
+    call_request::CallRequest, public_data_update_request::PublicDataUpdateRequest,
+    side_effect::{SideEffect, SideEffectLinkedToNoteHash}
 }
 };
 use crate::constants::{
-    MAX_NOTE_HASH_READ_REQUESTS_PER_TX, MAX_NULLIFIER_READ_REQUESTS_PER_TX,
-    MAX_NULLIFIER_KEY_VALIDATION_REQUESTS_PER_TX, MAX_NEW_NOTE_HASHES_PER_TX, MAX_NEW_NULLIFIERS_PER_TX,
-    MAX_PRIVATE_CALL_STACK_LENGTH_PER_TX, MAX_PUBLIC_CALL_STACK_LENGTH_PER_TX,
-    MAX_NEW_L2_TO_L1_MSGS_PER_TX, MAX_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX, MAX_PUBLIC_DATA_READS_PER_TX,
-    NUM_FIELDS_PER_SHA256, MAX_NON_REVERTIBLE_NOTE_HASHES_PER_TX, MAX_NON_REVERTIBLE_NULLIFIERS_PER_TX,
-    MAX_NON_REVERTIBLE_PUBLIC_CALL_STACK_LENGTH_PER_TX, MAX_REVERTIBLE_NOTE_HASHES_PER_TX,
-    MAX_REVERTIBLE_NULLIFIERS_PER_TX, MAX_REVERTIBLE_PUBLIC_CALL_STACK_LENGTH_PER_TX,
-    MAX_REVERTIBLE_PUBLIC_DATA_READS_PER_TX, MAX_REVERTIBLE_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX,
-    MAX_NON_REVERTIBLE_PUBLIC_DATA_READS_PER_TX, MAX_NON_REVERTIBLE_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX
+    MAX_NEW_NOTE_HASHES_PER_TX, MAX_NEW_NULLIFIERS_PER_TX, MAX_PRIVATE_CALL_STACK_LENGTH_PER_TX,
+    MAX_PUBLIC_CALL_STACK_LENGTH_PER_TX, MAX_NEW_L2_TO_L1_MSGS_PER_TX,
+    MAX_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX, NUM_FIELDS_PER_SHA256
 };
 
 use dep::std::unsafe;
@@ -32,10 +25,6 @@ use crate::traits::is_empty;
 use crate::utils::arrays::{array_cp, array_concat, array_to_bounded_vec};
 
 struct CombinedAccumulatedDataBuilder {
-    note_hash_read_requests: BoundedVec<SideEffect,MAX_NOTE_HASH_READ_REQUESTS_PER_TX>,
-    nullifier_read_requests: BoundedVec<ReadRequestContext, MAX_NULLIFIER_READ_REQUESTS_PER_TX>,
-    nullifier_key_validation_requests: BoundedVec<NullifierKeyValidationRequestContext, MAX_NULLIFIER_KEY_VALIDATION_REQUESTS_PER_TX>,
-
     new_note_hashes: BoundedVec<SideEffect, MAX_NEW_NOTE_HASHES_PER_TX>,
     new_nullifiers: BoundedVec<SideEffectLinkedToNoteHash, MAX_NEW_NULLIFIERS_PER_TX>,
 
@@ -52,8 +41,6 @@ struct CombinedAccumulatedDataBuilder {
     unencrypted_log_preimages_length: Field,
 
     public_data_update_requests: BoundedVec<PublicDataUpdateRequest, MAX_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX>,
-
-    public_data_reads: BoundedVec<PublicDataRead, MAX_PUBLIC_DATA_READS_PER_TX>,
 }
 
 impl CombinedAccumulatedDataBuilder {
@@ -62,9 +49,6 @@ impl CombinedAccumulatedDataBuilder {
         revertible: PublicAccumulatedRevertibleData
     ) -> CombinedAccumulatedDataBuilder {
         CombinedAccumulatedDataBuilder {
-            note_hash_read_requests: array_to_bounded_vec(revertible.note_hash_read_requests),
-            nullifier_read_requests: array_to_bounded_vec(revertible.nullifier_read_requests),
-            nullifier_key_validation_requests: array_to_bounded_vec(revertible.nullifier_key_validation_requests),
             new_note_hashes: array_to_bounded_vec(array_concat(non_revertible.new_note_hashes, revertible.new_note_hashes)),
             new_nullifiers: array_to_bounded_vec(array_concat(non_revertible.new_nullifiers, revertible.new_nullifiers)),
             private_call_stack: array_to_bounded_vec(revertible.private_call_stack),
@@ -84,21 +68,12 @@ impl CombinedAccumulatedDataBuilder {
                     non_revertible.public_data_update_requests,
                     revertible.public_data_update_requests
                 )
-            ),
-            public_data_reads: array_to_bounded_vec(
-                array_concat(
-                    non_revertible.public_data_reads,
-                    revertible.public_data_reads
-                )
             )
         }
     }
 
     pub fn finish(self) -> CombinedAccumulatedData {
         CombinedAccumulatedData {
-            note_hash_read_requests: self.note_hash_read_requests.storage,
-            nullifier_read_requests: self.nullifier_read_requests.storage,
-            nullifier_key_validation_requests: self.nullifier_key_validation_requests.storage,
             new_note_hashes: self.new_note_hashes.storage,
             new_nullifiers: self.new_nullifiers.storage,
             private_call_stack: self.private_call_stack.storage,
@@ -108,8 +83,7 @@ impl CombinedAccumulatedDataBuilder {
             unencrypted_logs_hash: self.unencrypted_logs_hash,
             encrypted_log_preimages_length: self.encrypted_log_preimages_length,
             unencrypted_log_preimages_length: self.unencrypted_log_preimages_length,
-            public_data_update_requests: self.public_data_update_requests.storage,
-            public_data_reads: self.public_data_reads.storage
+            public_data_update_requests: self.public_data_update_requests.storage
         }
     }
 
@@ -129,9 +103,6 @@ impl CombinedAccumulatedDataBuilder {
 
     pub fn to_public_accumulated_revertible_data(self) -> PublicAccumulatedRevertibleData {
         PublicAccumulatedRevertibleData {
-            note_hash_read_requests: self.note_hash_read_requests.storage,
-            nullifier_read_requests: self.nullifier_read_requests.storage,
-            nullifier_key_validation_requests: self.nullifier_key_validation_requests.storage,
             new_note_hashes: array_cp(self.new_note_hashes.storage),
             new_nullifiers: array_cp(self.new_nullifiers.storage),
             private_call_stack: self.private_call_stack.storage,
@@ -141,8 +112,7 @@ impl CombinedAccumulatedDataBuilder {
             unencrypted_logs_hash: self.unencrypted_logs_hash,
             encrypted_log_preimages_length: self.encrypted_log_preimages_length,
             unencrypted_log_preimages_length: self.unencrypted_log_preimages_length,
-            public_data_update_requests: array_cp(self.public_data_update_requests.storage),
-            public_data_reads: array_cp(self.public_data_reads.storage)
+            public_data_update_requests: array_cp(self.public_data_update_requests.storage)
         }
     }
