+++ noir-projects/noir-protocol-circuits/crates/types/src/abis/kernel_circuit_public_inputs/private_kernel_circuit_public_inputs_builder.nr
@@ -3,7 +3,8 @@ use crate::abis::{
     kernel_circuit_public_inputs::{
     private_kernel_inner_circuit_public_inputs::PrivateKernelInnerCircuitPublicInputs,
     private_kernel_tail_circuit_public_inputs::PrivateKernelTailCircuitPublicInputs
-}
+},
+    validation_requests::validation_requests_builder::ValidationRequestsBuilder
 };
 
 use crate::mocked::AggregationObject;
@@ -11,6 +12,7 @@ use crate::mocked::AggregationObject;
 struct PrivateKernelCircuitPublicInputsBuilder {
     aggregation_object: AggregationObject,
     min_revertible_side_effect_counter: u32,
+    validation_requests: ValidationRequestsBuilder,
     end: CombinedAccumulatedDataBuilder,
     constants: CombinedConstantData,
     is_private: bool,
@@ -21,6 +23,7 @@ impl PrivateKernelCircuitPublicInputsBuilder {
         PrivateKernelInnerCircuitPublicInputs {
             aggregation_object: self.aggregation_object,
             min_revertible_side_effect_counter: self.min_revertible_side_effect_counter,
+            validation_requests: self.validation_requests.finish(),
             end: self.end.finish(),
             constants: self.constants,
             is_private: self.is_private
