+++ noir-projects/noir-protocol-circuits/crates/types/src/abis/kernel_circuit_public_inputs/public_kernel_circuit_public_inputs.nr
@@ -1,12 +1,12 @@
 use crate::abis::{
     accumulated_data::{PublicAccumulatedNonRevertibleData, PublicAccumulatedRevertibleData},
-    combined_constant_data::CombinedConstantData
+    combined_constant_data::CombinedConstantData, validation_requests::ValidationRequests
 };
-
 use crate::mocked::AggregationObject;
 
 struct PublicKernelCircuitPublicInputs {
     aggregation_object: AggregationObject,
+    validation_requests: ValidationRequests,
     end_non_revertible: PublicAccumulatedNonRevertibleData,
     end: PublicAccumulatedRevertibleData,
     constants: CombinedConstantData,
