+++ noir-projects/noir-protocol-circuits/crates/types/src/abis/validation_requests.nr
@@ -1,0 +1,5 @@
+mod validation_requests;
+mod validation_requests_builder;
+
+use validation_requests::ValidationRequests;
+use validation_requests_builder::ValidationRequestsBuilder;
