+++ noir-projects/noir-protocol-circuits/crates/types/src/constants.nr
@@ -55,8 +55,6 @@ global MAX_NON_REVERTIBLE_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX: u64 = 16;
 global MAX_REVERTIBLE_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX: u64 = 16;
 
 global MAX_PUBLIC_DATA_READS_PER_TX: u64 = 32;
-global MAX_NON_REVERTIBLE_PUBLIC_DATA_READS_PER_TX: u64 = 16;
-global MAX_REVERTIBLE_PUBLIC_DATA_READS_PER_TX: u64 = 16;
 
 global MAX_NEW_L2_TO_L1_MSGS_PER_TX: u64 = 2;
 global MAX_NOTE_HASH_READ_REQUESTS_PER_TX: u64 = 128;
