+++ yarn-project/circuits.js/src/constants.gen.ts
@@ -27,8 +27,6 @@ export const MAX_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX = 32;
 export const MAX_NON_REVERTIBLE_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX = 16;
 export const MAX_REVERTIBLE_PUBLIC_DATA_UPDATE_REQUESTS_PER_TX = 16;
 export const MAX_PUBLIC_DATA_READS_PER_TX = 32;
-export const MAX_NON_REVERTIBLE_PUBLIC_DATA_READS_PER_TX = 16;
-export const MAX_REVERTIBLE_PUBLIC_DATA_READS_PER_TX = 16;
 export const MAX_NEW_L2_TO_L1_MSGS_PER_TX = 2;
 export const MAX_NOTE_HASH_READ_REQUESTS_PER_TX = 128;
 export const MAX_NULLIFIER_READ_REQUESTS_PER_TX = 8;
