+++ yarn-project/circuits.js/src/structs/kernel/public_kernel_circuit_public_inputs.ts
@@ -3,6 +3,7 @@ import { BufferReader, serializeToBuffer } from '@aztec/foundation/serialize';
 import { inspect } from 'util';
 
 import { AggregationObject } from '../aggregation_object.js';
+import { ValidationRequests } from '../validation_requests.js';
 import {
   CombinedAccumulatedData,
   PublicAccumulatedNonRevertibleData,
@@ -22,6 +23,10 @@ export class PublicKernelCircuitPublicInputs {
      * Aggregated proof of all the previous kernel iterations.
      */
     public aggregationObject: AggregationObject, // Contains the aggregated proof of all previous kernel iterations
+    /**
+     * Validation requests accumulated from public functions.
+     */
+    public validationRequests: ValidationRequests,
     /**
      * Accumulated side effects and enqueued calls that are not revertible.
      */
@@ -55,6 +60,7 @@ export class PublicKernelCircuitPublicInputs {
   toBuffer() {
     return serializeToBuffer(
       this.aggregationObject,
+      this.validationRequests,
       this.endNonRevertibleData,
       this.end,
       this.constants,
@@ -85,6 +91,7 @@ export class PublicKernelCircuitPublicInputs {
     const reader = BufferReader.asReader(buffer);
     return new PublicKernelCircuitPublicInputs(
       reader.readObject(AggregationObject),
+      reader.readObject(ValidationRequests),
       reader.readObject(PublicAccumulatedNonRevertibleData),
       reader.readObject(PublicAccumulatedRevertibleData),
       reader.readObject(CombinedConstantData),
@@ -98,6 +105,7 @@ export class PublicKernelCircuitPublicInputs {
   static empty() {
     return new PublicKernelCircuitPublicInputs(
       AggregationObject.makeFake(),
+      ValidationRequests.empty(),
       PublicAccumulatedNonRevertibleData.empty(),
       PublicAccumulatedRevertibleData.empty(),
       CombinedConstantData.empty(),
@@ -111,6 +119,7 @@ export class PublicKernelCircuitPublicInputs {
   [inspect.custom]() {
     return `PublicKernelCircuitPublicInputs {
       aggregationObject: ${this.aggregationObject},
+      validationRequests: ${inspect(this.validationRequests)},
       endNonRevertibleData: ${inspect(this.endNonRevertibleData)},
       end: ${inspect(this.end)},
       constants: ${this.constants},
