+++ yarn-project/circuits.js/src/structs/public_data_read_request.ts
@@ -1,0 +1,61 @@
+import { Fr } from '@aztec/foundation/fields';
+import { BufferReader, serializeToBuffer } from '@aztec/foundation/serialize';
+
+// TODO: Rename to PublicDataReadRequest
+/**
+ * Read operations from the public state tree.
+ */
+export class PublicDataRead {
+  constructor(
+    /**
+     * Index of the leaf in the public data tree.
+     */
+    public readonly leafSlot: Fr,
+    /**
+     * Returned value from the public data tree.
+     */
+    public readonly value: Fr,
+    /**
+     * Optional side effect counter tracking position of this event in tx execution.
+     */
+    public readonly sideEffectCounter?: number,
+  ) {}
+
+  static from(args: {
+    /**
+     * Index of the leaf in the public data tree.
+     */
+    leafIndex: Fr;
+    /**
+     * Returned value from the public data tree.
+     */
+    value: Fr;
+  }) {
+    return new PublicDataRead(args.leafIndex, args.value);
+  }
+
+  toBuffer() {
+    return serializeToBuffer(this.leafSlot, this.value);
+  }
+
+  isEmpty() {
+    return this.leafSlot.isZero() && this.value.isZero();
+  }
+
+  static fromBuffer(buffer: Buffer | BufferReader) {
+    const reader = BufferReader.asReader(buffer);
+    return new PublicDataRead(Fr.fromBuffer(reader), Fr.fromBuffer(reader));
+  }
+
+  static empty() {
+    return new PublicDataRead(Fr.ZERO, Fr.ZERO);
+  }
+
+  toFriendlyJSON() {
+    return `Leaf=${this.leafSlot.toFriendlyJSON()}: ${this.value.toFriendlyJSON()}`;
+  }
+
+  equals(other: PublicDataRead) {
+    return this.leafSlot.equals(other.leafSlot) && this.value.equals(other.value);
+  }
+}
